The .pypirc file is read from docker-context-files (#11779)
If you used context from git repo, the .piprc file was missing and
COPY in Dockerfile is not conditional.

This change copies the .pypirc conditionally from the
docker-context-files folder instead.

Also it was needlessly copied in the main image where it is not
needed and it was even dangerous to do so.

(cherry picked from commit 0a998f3aabaf26251250ac5cfd9745260e5c502f)